[강운덕] [LUCYSUS-1744] spanId, parentSpanId를 int->long으로 다시 변경함.
git-svn-id: http://svn.bds.nhncorp.com/pe/hippo-profiler-bootstrap/trunk@2731 84d0f5b1-2673-498c-a247-62c4ff18d310

+++ src/main/java/com/nhn/pinpoint/profiler/context/Trace.java
@@ -75,7 +75,7 @@ public interface Trace {
 
     void recordRemoteAddress(String remoteAddress);
 
-    void recordNextSpanId(int spanId);
+    void recordNextSpanId(long spanId);
 
     void recordParentApplication(String parentApplicationName, short parentApplicationType);
